requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/test.in
#
atomicwrites==1.4.1
    # via pytest
attrs==22.1.0
    # via pytest
click==8.1.3
    # via -r requirements\production.txt
colorama==0.4.5
    # via
    #   -r requirements\production.txt
    #   click
    #   pytest
commonmark==0.9.1
    # via rich
coverage[toml]==6.4.4
    # via pytest-cov
importlib-metadata==4.12.0
    # via pytest-randomly
iniconfig==1.1.1
    # via pytest
packaging==21.3
    # via
    #   pytest
    #   pytest-sugar
pluggy==1.0.0
    # via pytest
pprintpp==0.4.0
    # via pytest-clarity
py==1.11.0
    # via pytest
pygments==2.13.0
    # via rich
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements/test.in
    #   pytest-clarity
    #   pytest-cov
    #   pytest-randomly
    #   pytest-sugar
    #   pytest-tldr
pytest-clarity==1.0.1
    # via -r requirements/test.in
pytest-cov==3.0.0
    # via -r requirements/test.in
pytest-randomly==3.12.0
    # via -r requirements/test.in
pytest-sugar==0.9.5
    # via -r requirements/test.in
pytest-tldr==0.2.4
    # via -r requirements/test.in
rich==12.5.1
    # via pytest-clarity
termcolor==1.1.0
    # via pytest-sugar
tomli==2.0.1
    # via
    #   coverage
    #   pytest
zipp==3.8.1
    # via importlib-metadata
